play animations on whole group

+++ src/Endpoint/ALED/Animation/Mask/Slide.hs
@@ -7,7 +7,7 @@ import           Ivory.Language
 import           Ivory.Stdlib
 
 
-renderSlide :: Uint8
+renderSlide :: Uint16
             -> Sint32
             -> Record AnimationStruct
             -> Ivory (AllowBreak (ProcEffects s ())) IFloat
